+++ src/main/java/com/github/romualdrousseau/shuju/cv/Filter.java
@@ -8,7 +8,7 @@ public class Filter {
     public void apply(ISearchBitmap searchBitmap, double threshold) {
         for (int y = 0; y < searchBitmap.getHeight(); y++) {
             for (int x = 0; x < searchBitmap.getWidth(); x++) {
-                int acc = this.filter.sobel(searchBitmap, x, y);
+                float acc = this.filter.sobel(searchBitmap, x, y);
                 if (acc < threshold) {
                     searchBitmap.set(x, y, 0);
                 }
@@ -19,7 +19,7 @@ public class Filter {
     public void apply(ISearchBitmap searchBitmap, int[] clip, double threshold) {
         for (int y = clip[1]; y < clip[3]; y++) {
             for (int x = clip[0]; x < clip[2]; x++) {
-                int acc = this.filter.sobel(searchBitmap, x, y);
+                float acc = this.filter.sobel(searchBitmap, x, y);
                 if (acc < threshold) {
                     searchBitmap.set(x, y, 0);
                 }
@@ -30,7 +30,7 @@ public class Filter {
     public void applyNeg(ISearchBitmap searchBitmap, double threshold) {
         for (int y = 0; y < searchBitmap.getHeight(); y++) {
             for (int x = 0; x < searchBitmap.getWidth(); x++) {
-                int acc = this.filter.sobel(searchBitmap, x, y);
+                float acc = this.filter.sobel(searchBitmap, x, y);
                 if (acc >= threshold) {
                     searchBitmap.set(x, y, 1);
                 }
@@ -41,7 +41,7 @@ public class Filter {
     public void applyNeg(ISearchBitmap searchBitmap, int[] clip, double threshold) {
         for (int y = clip[1]; y < clip[3]; y++) {
             for (int x = clip[0]; x < clip[2]; x++) {
-                int acc = this.filter.sobel(searchBitmap, x, y);
+                float acc = this.filter.sobel(searchBitmap, x, y);
                 if (acc >= threshold) {
                     searchBitmap.set(x, y, 1);
                 }
@@ -52,7 +52,7 @@ public class Filter {
     public void apply(ISearchBitmap sourceBitmap, ISearchBitmap destBitmap, double threshold) {
         for (int y = 0; y < sourceBitmap.getHeight(); y++) {
             for (int x = 0; x < sourceBitmap.getWidth(); x++) {
-                int acc = this.filter.sobel(sourceBitmap, x, y);
+                float acc = this.filter.sobel(sourceBitmap, x, y);
                 if (acc < threshold) {
                     destBitmap.set(x, y, 0);
                 } else {
@@ -65,7 +65,7 @@ public class Filter {
     public void apply(ISearchBitmap sourceBitmap, ISearchBitmap destBitmap, int[] clip, double threshold) {
         for (int y = clip[1]; y < clip[3]; y++) {
             for (int x = clip[0]; x < clip[2]; x++) {
-                int acc = this.filter.sobel(sourceBitmap, x, y);
+                float acc = this.filter.sobel(sourceBitmap, x, y);
                 if (acc < threshold) {
                     destBitmap.set(x, y, 0);
                 } else {
